managers

+++ src/main/java/com/roi/dao/RoleDao.java
@@ -3,6 +3,6 @@ import com.roi.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RoleDao extends JpaRepository<Role, Long>{
+public interface RoleDao extends JpaRepository<Role, Integer>{
 
 }
